the Best-first_algorithm

+++ ISearcher.h
@@ -14,7 +14,7 @@ public:
 
     virtual solution search(ISearchable<T> *searchable) = 0;
 
-    virtual int getNumberOfNodesEvaluated() = 0;
+    virtual unsigned long getNumberOfNodesEvaluated() = 0;
 };
